Module bonsaidb_local::config

Configuration options.

Structs

ArgonConfiguration

Password hashing configuration.

KeyValuePersistence

Rules for persisting key-value changes. Default persistence is to immediately persist all changes. While this ensures data integrity, the overhead of the key-value store can be significantly reduced by utilizing lazy persistence strategies that delay writing changes until certain thresholds have been met.

PersistenceThreshold

A threshold controlling lazy commits. For a threshold to apply, both number_of_changes must be met or surpassed and duration must have elpased since the last commit.

StorageConfiguration

Configuration options for Storage.

Tasks

Configuration options for background tasks.

TimedArgonParams

Automatic configuration based on execution time. This is measured during the first set_password

Views

Configuration options for views.

Enums

ArgonParams

Argon2id base parameters.

Compression

All available compression algorithms.

Traits

Builder

Storage configuration builder methods.