Trait bonsaidb_local::core::kv::Kv

pub trait Kv: Send + Sync {
    #[must_use]
    pub fn execute_key_operation<'life0, 'async_trait>(
        &'life0 self, 
        op: KeyOperation
    ) -> Pin<Box<dyn Future<Output = Result<Output, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    pub fn set_key<S, V>(&'a self, key: S, value: &'a V) -> Builder<'a, Self, V>
    where
        S: Into<String>,
        V: Serialize + Send + Sync,
    { ... }

    pub fn set_numeric_key<S, V>(
        &self, 
        key: S, 
        value: V
    ) -> Builder<'_, Self, ()>
    where
        S: Into<String>,
        V: Into<Numeric>,
    { ... }

    pub fn increment_key_by<S, V>(
        &self, 
        key: S, 
        value: V
    ) -> Builder<'_, Self, V>
    where
        S: Into<String> + Send + Sync,
        V: Into<Numeric> + TryFrom<Numeric, Error = IncompatibleTypeError> + Send + Sync,
    { ... }

    pub fn decrement_key_by<S, V>(
        &self, 
        key: S, 
        value: V
    ) -> Builder<'_, Self, V>
    where
        S: Into<String> + Send + Sync,
        V: Into<Numeric> + TryFrom<Numeric, Error = IncompatibleTypeError> + Send + Sync,
    { ... }

    pub fn get_key<S>(&self, key: S) -> Builder<'_, Self>
    where
        S: Into<String>,
    { ... }

    #[must_use]
    pub fn delete_key<'life0, 'async_trait, S>(
        &'life0 self, 
        key: S
    ) -> Pin<Box<dyn Future<Output = Result<KeyStatus, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
        S: Into<String> + Send + 'async_trait,
    { ... }

    pub fn key_namespace(&self) -> Option<&str> { ... }

    pub fn with_key_namespace(&self, namespace: &str) -> Namespaced<'_, Self> { ... }
}

Key-Value store methods. The Key-Value store is designed to be a high-performance, lightweight storage mechanism.

When compared to Collections, the Key-Value store does not offer ACID-compliant transactions. Instead, the Key-Value store is made more efficient by periodically flushing the store to disk rather than during each operation. As such, the Key-Value store is intended to be used as a lightweight caching layer. However, because each of the operations it supports are executed atomically, the Key-Value store can also be utilized for synchronized locking.

Required methods

#[must_use]
pub fn execute_key_operation<'life0, 'async_trait>(
&'life0 self,
op: KeyOperation
) -> Pin<Box<dyn Future<Output = Result<Output, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Executes a single KeyOperation.

Loading content...
Provided methods

pub fn set_key<S, V>(&'a self, key: S, value: &'a V) -> Builder<'a, Self, V> where
S: Into<String>,
V: Serialize + Send + Sync,

Sets key to value. This function returns a builder that is also a Future. Awaiting the builder will execute Command::Set with the options given.

pub fn set_numeric_key<S, V>(&self, key: S, value: V) -> Builder<'_, Self, ()> where
S: Into<String>,
V: Into<Numeric>,

Sets key to value. This stores the value as a Numeric, enabling atomic math operations to be performed on this key. This function returns a builder that is also a Future. Awaiting the builder will execute Command::Set with the options given.

pub fn increment_key_by<S, V>(&self, key: S, value: V) -> Builder<'_, Self, V> where
S: Into<String> + Send + Sync,
V: Into<Numeric> + TryFrom<Numeric, Error = IncompatibleTypeError> + Send + Sync,

Increments key by value. The value stored must be a Numeric, otherwise an error will be returned. The result of the increment will be the value’s type. For example, if the stored value is currently a u64, but value is a f64, the current value will be converted to an f64, and the stored value will be an f64.

pub fn decrement_key_by<S, V>(&self, key: S, value: V) -> Builder<'_, Self, V> where
S: Into<String> + Send + Sync,
V: Into<Numeric> + TryFrom<Numeric, Error = IncompatibleTypeError> + Send + Sync,

Decrements key by value. The value stored must be a Numeric, otherwise an error will be returned. The result of the decrement will be the value’s type. For example, if the stored value is currently a u64, but value is a f64, the current value will be converted to an f64, and the stored value will be an f64.

pub fn get_key<S>(&self, key: S) -> Builder<'_, Self> where
S: Into<String>,

Gets the value stored at key. This function returns a builder that is also a Future. Awaiting the builder will execute Command::Get with the options given.

#[must_use]
pub fn delete_key<'life0, 'async_trait, S>(
&'life0 self,
key: S
) -> Pin<Box<dyn Future<Output = Result<KeyStatus, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,
S: Into<String> + Send + 'async_trait,

Deletes the value stored at key.

pub fn key_namespace(&self) -> Option<&str>

The current namespace.

pub fn with_key_namespace(&self, namespace: &str) -> Namespaced<'_, Self>

Access this Key-Value store within a namespace. When using the returned Namespaced instance, all keys specified will be separated into their own storage designated by namespace.

Loading content...