Enum bonsaidb_local::core::networking::ServerRequest

pub enum ServerRequest {
    LoginWithPassword {
        username: String,
        login_request: LoginRequest,
    },
    FinishPasswordLogin {
        login_finalization: LoginFinalization,
    },
    CreateDatabase {
        database: Database,
        only_if_needed: bool,
    },
    DeleteDatabase {
        name: String,
    },
    ListDatabases,
    ListAvailableSchemas,
    CreateUser {
        username: String,
    },
    SetPassword {
        user: NamedReference<'static>,
        password_request: RegistrationRequest,
    },
    FinishSetPassword {
        user: NamedReference<'static>,
        password_finalization: RegistrationFinalization,
    },
    AlterUserPermissionGroupMembership {
        user: NamedReference<'static>,
        group: NamedReference<'static>,
        should_be_member: bool,
    },
    AlterUserRoleMembership {
        user: NamedReference<'static>,
        role: NamedReference<'static>,
        should_be_member: bool,
    },
}

A server-related request.

Variants

LoginWithPassword

Authenticates the current session as username using a password.

Fields of LoginWithPassword

username: String

The username of the user to authenticate as.

login_request: LoginRequest

The password login request.

FinishPasswordLogin

Completes logging in via LoginWithPassword.

Fields of FinishPasswordLogin

login_finalization: LoginFinalization

The payload required to complete logging in.

CreateDatabase

Creates a database.

Fields of CreateDatabase

database: Database

The database to create.

only_if_needed: bool

Only attempts to create the database if it doesn’t already exist.

DeleteDatabase

Deletes the database named name

Fields of DeleteDatabase

name: String

The name of the database to delete.

ListDatabases

Lists all databases.

ListAvailableSchemas

Lists available schemas.

CreateUser

Creates a user.

Fields of CreateUser

username: String

The unique username of the user to create.

SetPassword

Sets a user’s password.

Fields of SetPassword

user: NamedReference<'static>

The username or id of the user to set the password for.

password_request: RegistrationRequest

A registration request for a password.

FinishSetPassword

Finishes setting the password for a user.

Fields of FinishSetPassword

user: NamedReference<'static>

The username or id of the user to set the password for.

password_finalization: RegistrationFinalization

The finalization payload for the password change.

AlterUserPermissionGroupMembership

Alter’s a user’s membership in a permission group.

Fields of AlterUserPermissionGroupMembership

user: NamedReference<'static>

The username or id of the user.

group: NamedReference<'static>

The name or id of the group.

should_be_member: bool

Whether the user should be in the group.

AlterUserRoleMembership

Alter’s a user’s role

Fields of AlterUserRoleMembership

user: NamedReference<'static>

The username or id of the user.

role: NamedReference<'static>

The name or id of the role.

should_be_member: bool

Whether the user should have the role.