Trait bonsaidb_local::core::schema::Collection

pub trait Collection: Debug + Send + Sync {
    pub fn collection_name() -> Result<CollectionName, InvalidNameError>;

    pub fn define_views(schema: &mut Schematic) -> Result<(), Error>;

    #[must_use]
    pub fn default_encryption_key() -> Option<KeyId> { ... }

    #[must_use]
    pub fn get<'life0, 'async_trait, C>(
        id: u64, 
        connection: &'life0 C
    ) -> Pin<Box<dyn Future<Output = Result<Option<CollectionDocument<Self>>, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: Serialize + for<'de> Deserialize<'de> + 'async_trait,
        C: Connection + 'async_trait,
    { ... }

    #[must_use]
    pub fn insert_into<'life0, 'async_trait, Cn>(
        self, 
        connection: &'life0 Cn
    ) -> Pin<Box<dyn Future<Output = Result<CollectionDocument<Self>, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: Serialize + for<'de> Deserialize<'de> + 'static + 'async_trait,
        Cn: Connection + 'async_trait,
    { ... }
}

A namespaced collection of Document<Self> items and views.

Required methods

pub fn collection_name() -> Result<CollectionName, InvalidNameError>

The Id of this collection.

pub fn define_views(schema: &mut Schematic) -> Result<(), Error>

Defines all Views in this collection in schema.

Loading content...
Provided methods

#[must_use]
pub fn default_encryption_key() -> Option<KeyId>

If a KeyId is returned, documents will be encrypted with this key if no key is specified during creation.

#[must_use]
pub fn get<'life0, 'async_trait, C>(
id: u64,
connection: &'life0 C
) -> Pin<Box<dyn Future<Output = Result<Option<CollectionDocument<Self>>, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: Serialize + for<'de> Deserialize<'de> + 'async_trait,
C: Connection + 'async_trait,

Gets a CollectionDocument with id from connection.

#[must_use]
pub fn insert_into<'life0, 'async_trait, Cn>(
self,
connection: &'life0 Cn
) -> Pin<Box<dyn Future<Output = Result<CollectionDocument<Self>, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: Serialize + for<'de> Deserialize<'de> + 'static + 'async_trait,
Cn: Connection + 'async_trait,

Inserts this value into the collection, returning the created document.

Loading content...