Struct bonsaidb_local::AsyncDatabase

pub struct AsyncDatabase { /* private fields */ }

A database stored in BonsaiDb. This type is designed for use with Tokio. For blocking (non-asynchronous) code, see the Database type instead.

Converting between Async and Blocking Types

AsyncDatabase and Database can be converted to and from each other using:

AsyncDatabase::into_blocking()

AsyncDatabase::to_blocking()

AsyncDatabase::as_blocking()

Database::into_async()

Database::to_async()

Database::into_async_with_runtime()

Database::to_async_with_runtime()

Using Database to create a single database

Databaseprovides an easy mechanism to open and access a single database:

// `bonsaidb_core` is re-exported to `bonsaidb::core` or `bonsaidb_local::core`.
use bonsaidb_core::schema::Collection;
// `bonsaidb_local` is re-exported to `bonsaidb::local` if using the omnibus crate.
use bonsaidb_local::{
    config::{Builder, StorageConfiguration},
    AsyncDatabase,
};
use serde::{Deserialize, Serialize};

#[derive(Debug, Serialize, Deserialize, Collection)]
#[collection(name = "blog-posts")]
struct BlogPost {
    pub title: String,
    pub contents: String,
}

let db = AsyncDatabase::open::<BlogPost>(StorageConfiguration::new("my-db.bonsaidb")).await?;

Under the hood, this initializes a AsyncStorage instance pointing at “./my-db.bonsaidb”. It then returns (or creates) a database named “default” with the schema BlogPost.

In this example, BlogPost implements the Collection trait, and all collections can be used as a [Schema].

Implementations

impl AsyncDatabase

pub async fn open<DB: Schema>(
configuration: StorageConfiguration
) -> Result<Self, Error>

Creates a Storage with a single-database named “default” with its data stored at path.

pub fn with_effective_permissions(
&self,
effective_permissions: Permissions
) -> Option<Self>

Restricts an unauthenticated instance to having effective_permissions. Returns None if a session has already been established.

pub fn into_blocking(self) -> Database

Converts this instance into its blocking version, which is able to be used without async.

pub fn to_blocking(&self) -> Database

Converts this instance into its blocking version, which is able to be used without async.

pub fn as_blocking(&self) -> &Database

Returns a reference to this instance’s blocking version, which is able to be used without async.