Struct bonsaidb_local::Storage

pub struct Storage { /* fields omitted */ }

A file-based, multi-database, multi-user database engine.

Implementations

impl Storage

pub async fn open_local<P: AsRef<Path> + Send>(
path: P,
configuration: Configuration
) -> Result<Self, Error>

Creates or opens a multi-database Storage with its data stored in directory.

pub fn path(&self) -> &Path

Returns the path of the database storage.

pub fn unique_id(&self) -> StorageId

Returns the unique id of the server.

This value is set from the Configuration or randomly generated when creating a server. It shouldn’t be changed after a server is in use, as doing can cause issues. For example, the vault that manages encrypted storage uses the server ID to store the vault key. If the server ID changes, the vault key storage will need to be updated with the new server ID.

pub async fn register_schema<DB: Schema>(&self) -> Result<(), Error>

Registers a schema for use within the server.

pub async fn database<DB: Schema>(
&self,
name: &str
) -> Result<Database<DB>, Error>

Retrieves a database. This function checks that the database exists and that the schema matches.

pub async fn admin(&self) -> Database<Admin>

Returns the administration database.