Trait bonsaidb_local::BackupLocation

pub trait BackupLocation: Send + Sync {
    type Error: AnyError;
    fn store<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait>(
        &'life0 self, 
        schema: &'life1 SchemaName, 
        database_name: &'life2 str, 
        container: &'life3 str, 
        name: &'life4 str, 
        object: &'life5 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        'life5: 'async_trait,
        Self: 'async_trait;

    fn list_schemas<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SchemaName>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn list_databases<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        schema: &'life1 SchemaName
    ) -> Pin<Box<dyn Future<Output = Result<Vec<String>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn list_stored<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self, 
        schema: &'life1 SchemaName, 
        database_name: &'life2 str, 
        container: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<Vec<String>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait;

    fn load<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        schema: &'life1 SchemaName, 
        database_name: &'life2 str, 
        container: &'life3 str, 
        name: &'life4 str
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait;
}

A location to store and restore a database from.

Associated Types

type Error: AnyError

The error type for the backup location.

Required methods

fn store<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait>(
&'life0 self,
schema: &'life1 SchemaName,
database_name: &'life2 str,
container: &'life3 str,
name: &'life4 str,
object: &'life5 [u8]
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
'life5: 'async_trait,
Self: 'async_trait,

Store object at path with name.

fn list_schemas<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<SchemaName>, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Lists all of the schemas stored in this backup location.

fn list_databases<'life0, 'life1, 'async_trait>(
&'life0 self,
schema: &'life1 SchemaName
) -> Pin<Box<dyn Future<Output = Result<Vec<String>, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

List all of the names of the databases stored for schema.

fn list_stored<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
schema: &'life1 SchemaName,
database_name: &'life2 str,
container: &'life3 str
) -> Pin<Box<dyn Future<Output = Result<Vec<String>, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

List all stored named objects at path. The names should be the same that were provided when store() was called.

fn load<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
schema: &'life1 SchemaName,
database_name: &'life2 str,
container: &'life3 str,
name: &'life4 str
) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Load a previously stored object from path with name.