Struct bonsaidb_local::AsyncStorage

pub struct AsyncStorage { /* private fields */ }

A file-based, multi-database, multi-user database engine. This type is designed for use with Tokio. For blocking (non-asynchronous) code, see the Storage type instead.

Converting between Blocking and Async Types

AsyncDatabase and Database can be converted to and from each other using:

AsyncStorage::into_blocking()

AsyncStorage::to_blocking()

AsyncStorage::as_blocking()

Storage::into_async()

Storage::to_async()

Storage::into_async_with_runtime()

Storage::to_async_with_runtime()

Converting from AsyncDatabase::open to AsyncStorage::open

AsyncDatabase::open is a simple method that uses AsyncStorage to create a database named default with the schema provided. These two ways of opening the database are the same:

// `bonsaidb_core` is re-exported to `bonsaidb::core` or `bonsaidb_local::core`.
use bonsaidb_core::{connection::AsyncStorageConnection, schema::Schema};
// `bonsaidb_local` is re-exported to `bonsaidb::local` if using the omnibus crate.
use bonsaidb_local::{
    config::{Builder, StorageConfiguration},
    AsyncDatabase, AsyncStorage,
};
// This creates a Storage instance, creates a database, and returns it.
let db = AsyncDatabase::open::<MySchema>(StorageConfiguration::new("my-db.bonsaidb")).await?;

// This is the equivalent code being executed:
let storage =
    AsyncStorage::open(StorageConfiguration::new("my-db.bonsaidb").with_schema::<MySchema>()?)
        .await?;
storage.create_database::<MySchema>("default", true).await?;
let db = storage.database::<MySchema>("default").await?;

Using multiple databases

This example shows how to use AsyncStorage to create and use multiple databases with multiple schemas:

use bonsaidb_core::{
    connection::AsyncStorageConnection,
    schema::{Collection, Schema},
};
use bonsaidb_local::{
    config::{Builder, StorageConfiguration},
    AsyncStorage,
};
use serde::{Deserialize, Serialize};

#[derive(Debug, Schema)]
#[schema(name = "my-schema", collections = [BlogPost, Author])]
struct MySchema;

#[derive(Debug, Serialize, Deserialize, Collection)]
#[collection(name = "blog-posts")]
struct BlogPost {
    pub title: String,
    pub contents: String,
    pub author_id: u64,
}

#[derive(Debug, Serialize, Deserialize, Collection)]
#[collection(name = "blog-posts")]
struct Author {
    pub name: String,
}

let storage = AsyncStorage::open(
    StorageConfiguration::new("my-db.bonsaidb")
        .with_schema::<BlogPost>()?
        .with_schema::<MySchema>()?,
)
.await?;

storage
    .create_database::<BlogPost>("ectons-blog", true)
    .await?;
let ectons_blog = storage.database::<BlogPost>("ectons-blog").await?;
storage
    .create_database::<MySchema>("another-db", true)
    .await?;
let another_db = storage.database::<MySchema>("another-db").await?;

Implementations

impl AsyncStorage

pub async fn open(configuration: StorageConfiguration) -> Result<Self, Error>

Creates or opens a multi-database AsyncStorage with its data stored in directory.

pub async fn restore<L: AnyBackupLocation + 'static>(
&self,
location: L
) -> Result<(), Error>

Restores all data from a previously stored backup location.

pub async fn backup<L: AnyBackupLocation + 'static>(
&self,
location: L
) -> Result<(), Error>

Stores a copy of all data in this instance to location.

pub fn with_effective_permissions(
&self,
effective_permissions: Permissions
) -> Option<Self>

Restricts an unauthenticated instance to having effective_permissions. Returns None if a session has already been established.

pub fn into_blocking(self) -> Storage

Converts this instance into its blocking version, which is able to be used without async.

pub fn to_blocking(&self) -> Storage

Converts this instance into its blocking version, which is able to be used without async.

pub fn as_blocking(&self) -> &Storage

Returns a reference to this instance’s blocking version, which is able to be used without async.