Trait bonsaidb_local::BackupLocation

pub trait BackupLocation: Send + Sync {
    type Error: AnyError;

    fn store(
        &self,
        schema: &SchemaName,
        database_name: &str,
        container: &str,
        name: &str,
        object: &[u8]
    ) -> Result<(), Self::Error>;
    fn list_schemas(&self) -> Result<Vec<SchemaName>, Self::Error>;
    fn list_databases(
        &self,
        schema: &SchemaName
    ) -> Result<Vec<String>, Self::Error>;
    fn list_stored(
        &self,
        schema: &SchemaName,
        database_name: &str,
        container: &str
    ) -> Result<Vec<String>, Self::Error>;
    fn load(
        &self,
        schema: &SchemaName,
        database_name: &str,
        container: &str,
        name: &str
    ) -> Result<Vec<u8>, Self::Error>;
}

A location to store and restore a database from.

Required Associated Types

type Error: AnyError

The error type for the backup location.

Required Methods

fn store(
&self,
schema: &SchemaName,
database_name: &str,
container: &str,
name: &str,
object: &[u8]
) -> Result<(), Self::Error>

Store object at path with name.

fn list_schemas(&self) -> Result<Vec<SchemaName>, Self::Error>

Lists all of the schemas stored in this backup location.

fn list_databases(&self, schema: &SchemaName) -> Result<Vec<String>, Self::Error>

List all of the names of the databases stored for schema.

fn list_stored(
&self,
schema: &SchemaName,
database_name: &str,
container: &str
) -> Result<Vec<String>, Self::Error>

List all stored named objects at path. The names should be the same that were provided when store() was called.

fn load(
&self,
schema: &SchemaName,
database_name: &str,
container: &str,
name: &str
) -> Result<Vec<u8>, Self::Error>

Load a previously stored object from path with name.