Trait bonsaidb_local::StorageNonBlocking

pub trait StorageNonBlocking: Sized {
    fn path(&self) -> &Path;
    fn assume_session(&self, session: Session) -> Result<Self, Error>;
}

Functionality that is available on both Storage and AsyncStorage.

Required Methods

fn path(&self) -> &Path

Returns the path of the database storage.

fn assume_session(&self, session: Session) -> Result<Self, Error>

Returns a new instance of Storage with session as the effective authentication session. This call will only succeed if there is no current session.